+++ packages/components/web/index.tsx
@@ -15,3 +15,4 @@ export * from './src/components/Switch/Switch';
 export * from './src/components/SwitchWithLabel/SwitchWithLabel';
 export * from './src/components/Rating/Rating';
 export * from './src/components/Autocomplete/Autocomplete';
+export * from './src/components/FloatingActionButton/FloatingActionButton';