+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 - SwitchWithLabel
 - Rating
 - Autocomplete
+- Floating Action Button
+
 
 ---